Add a shell authentication routine and rejection method.
Take the data from the login form and treat it as 'authorized' always
for now.

Pass along authenticated user in authentication module, for use in
controller views.

Move socket connections to home controller, so we don't connect
until the persons logged in (auth'd)

Next up is to actually wait for the server on a login, so we can
associate a user server side with a socket.

Then on their data is auth'd and passed as themselves.