Added support for getting general socket messages.
Handling of `UNAUTHORIZED` message for all sockets, to bump
the host to the login screen.

Need to add support for when the server disconnects itself from us.

Flesh out the chat window directive.  Add better support
for messages.  Change to showing messages in an ng-repeat.

Client now is responsible for asking to join a global chat.