Refactor chat window directive to now use a chat manager service.
This is a cleaner seperation of responsblity for the chat window,
and also allows us to have global messages and things like
private whispers more easily.

Rework some of the hacky side effects used to connect to a global chat
room in the home controller.

Add some more basic ui markup and template for the game states now
supported by the server.