Add support for the new game list socket flow.
Game controller now talks and listens to the game list as well,
and will connect to the game server when it gets the game started msg.

Fixed a bunch of bugs - a large bug in the authentication service
was not properly setting the servers login response.

Removed some code as well.