Add chat handler callbacks for chat window.
This lets the chat window directive give parent controllers access
to its scope methods in an angular safe way.

Initalize the web socket in the app's run state.

This lets our controllers just worry about connecting
when they need to.

Change how we expect data from the server to be formatted.  In general
the server now sends a `game` property and a `focus` property.